CR-1540 fixed test package name; improved unit test; replaced failure redaction code
