+++ src/main/java/uk/gov/ons/ctp/integration/ratelimiter/client/RateLimiterClient.java
@@ -26,6 +26,7 @@ import uk.gov.ons.ctp.integration.ratelimiter.model.LimitDescriptor;
 import uk.gov.ons.ctp.integration.ratelimiter.model.LimitStatus;
 import uk.gov.ons.ctp.integration.ratelimiter.model.RateLimitRequest;
 import uk.gov.ons.ctp.integration.ratelimiter.model.RateLimitResponse;
+import uk.gov.ons.ctp.integration.ratelimiter.util.Encryptor;
 
 public class RateLimiterClient {
   private static final Logger log = LoggerFactory.getLogger(RateLimiterClient.class);
@@ -147,7 +148,7 @@ public class RateLimiterClient {
         .with("caseType", caseType.name())
         .with("ipAddress", ipAddress)
         .with("uprn", uprn.getValue())
-        .with("telNo", telNo == null ? null : Encryptor.aesEncrypt(encryptionPassword, telNo))
+        .with("encrypted-telNo", encrypt(telNo))
         .info("Fulfilment rate limit. Going to call Rate Limiter Service");
 
     // Make it easy to access limiter parameters by adding to a hashmap
@@ -403,7 +404,7 @@ public class RateLimiterClient {
         // An expected failure scenario. Record the breach and make sure caller
         // knows by re-throwing the exception
         String breachDescription = describeLimitBreach(request, limiterException);
-        log.info(breachDescription.toString());
+        log.info(breachDescription);
         throw limiterException;
       } else {
         // Something unexpected went wrong
@@ -461,7 +462,8 @@ public class RateLimiterClient {
       String descriptorKey = descriptorEntry.getKey();
       String descriptorValue = descriptorEntry.getValue();
       if (descriptorKey.equals(DESC_TEL_NO)) {
-        descriptorValue = redactTelephoneNumber(descriptorValue);
+        descriptorKey = "encrypted-" + descriptorKey;
+        descriptorValue = encrypt(descriptorValue);
       }
 
       desc.append(descriptorKey + "=" + descriptorValue);
@@ -485,16 +487,7 @@ public class RateLimiterClient {
     return response;
   }
 
-  private String redactTelephoneNumber(String telNo) {
-    if (telNo == null) {
-      return "null";
-    }
-
-    StringBuilder redactedTelephoneNumber = new StringBuilder();
-
-    redactedTelephoneNumber.append("xxxx");
-    redactedTelephoneNumber.append(telNo.substring(telNo.length() - 2));
-
-    return redactedTelephoneNumber.toString();
+  private String encrypt(String telNo) {
+    return telNo == null ? null : Encryptor.aesEncrypt(encryptionPassword, telNo);
   }
 }
