+++ src/test/java/uk/gov/ons/ctp/integration/ratelimiter/client/RateLimiterClientFulfilmentTest.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.ctp.integration.ratelimiterclient;
+package uk.gov.ons.ctp.integration.ratelimiter.client;
 
 import static org.junit.Assert.assertThrows;
 import static org.junit.Assert.assertTrue;
@@ -10,15 +10,16 @@ import org.junit.runner.RunWith;
 import org.mockito.junit.MockitoJUnitRunner;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.server.ResponseStatusException;
+import uk.gov.ons.ctp.common.FixtureHelper;
 import uk.gov.ons.ctp.common.domain.CaseType;
 import uk.gov.ons.ctp.common.domain.UniquePropertyReferenceNumber;
 import uk.gov.ons.ctp.common.error.CTPException;
 import uk.gov.ons.ctp.integration.common.product.model.Product;
 import uk.gov.ons.ctp.integration.common.product.model.Product.DeliveryChannel;
 import uk.gov.ons.ctp.integration.common.product.model.Product.ProductGroup;
-import uk.gov.ons.ctp.integration.ratelimiter.client.RateLimiterClient;
 import uk.gov.ons.ctp.integration.ratelimiter.model.LimitDescriptor;
 import uk.gov.ons.ctp.integration.ratelimiter.model.RateLimitRequest;
+import uk.gov.ons.ctp.integration.ratelimiter.model.RateLimitResponse;
 
 /** This class contains unit tests for limit testing fulfilment requests. */
 @RunWith(MockitoJUnitRunner.class)
@@ -112,7 +113,8 @@ public class RateLimiterClientFulfilmentTest extends RateLimiterClientTestBase {
     // Limit request is going to fail with exception. This needs to contain a string with the
     // limiters
     // too-many-requests response
-    ResponseStatusException failureException = overTheLimitException();
+    RateLimitResponse resp = FixtureHelper.loadClassFixtures(RateLimitResponse[].class).get(0);
+    ResponseStatusException failureException = overTheLimitException(resp);
     mockRateLimitException(failureException);
 
     // Confirm that limiter request fails with a 429 exception
