+++ src/test/java/uk/gov/ons/ctp/integration/ratelimiter/client/RateLimiterClientWebformTest.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.ctp.integration.ratelimiterclient;
+package uk.gov.ons.ctp.integration.ratelimiter.client;
 
 import static org.junit.Assert.assertThrows;
 import static org.junit.Assert.assertTrue;
@@ -14,92 +14,61 @@ import uk.gov.ons.ctp.common.error.CTPException;
 import uk.gov.ons.ctp.integration.ratelimiter.model.LimitDescriptor;
 import uk.gov.ons.ctp.integration.ratelimiter.model.RateLimitRequest;
 
+/** This class contains unit tests for limit testing Webform requests. */
 @RunWith(MockitoJUnitRunner.class)
-public class RateLimiterClientEqLaunchTest extends RateLimiterClientTestBase {
+public class RateLimiterClientWebformTest extends RateLimiterClientTestBase {
 
   @Test
-  public void shouldRejectNullDomain() {
+  public void checkWebformRateLimit_nullDomain() {
     CTPException exception =
         assertThrows(
             CTPException.class,
-            () -> rateLimiterClient.checkEqLaunchLimit(null, AN_IPv4_ADDRESS, 10));
+            () -> rateLimiterClient.checkWebformRateLimit(null, AN_IPv4_ADDRESS));
     assertTrue(exception.getMessage(), exception.getMessage().contains("'domain' cannot be null"));
-    verifyEnvoyLimiterNotCalled();
-  }
-
-  @Test
-  public void shouldRejectZeroLoadSheddingModulus() {
-    CTPException exception =
-        assertThrows(
-            CTPException.class,
-            () -> rateLimiterClient.checkEqLaunchLimit(domain, AN_IPv4_ADDRESS, 0));
-    assertTrue(
-        exception.getMessage(),
-        exception.getMessage().contains("'loadSheddingModulus' cannot be zero"));
-    verifyEnvoyLimiterNotCalled();
   }
 
   @Test
   public void shouldQuietlyAcceptNullClientIpNotCallingLimiter() throws Exception {
-    rateLimiterClient.checkEqLaunchLimit(domain, null, 10);
+    rateLimiterClient.checkWebformRateLimit(domain, null);
     verifyEnvoyLimiterNotCalled();
   }
 
   @Test
   public void shouldQuietlyAcceptEmptyClientIpNotCallingLimiter() throws Exception {
-    rateLimiterClient.checkEqLaunchLimit(domain, "", 10);
+    rateLimiterClient.checkWebformRateLimit(domain, "");
     verifyEnvoyLimiterNotCalled();
   }
 
   @Test
   public void shouldQuietlyAcceptBadlyFormattedClientIpNotCallingLimiter() throws Exception {
-    rateLimiterClient.checkEqLaunchLimit(domain, "badlyformattedIpAddress", 10);
+    rateLimiterClient.checkWebformRateLimit(domain, "badlyformattedIpAddress");
     verifyEnvoyLimiterNotCalled();
   }
 
   @Test
   public void shouldQuietlyAcceptIpV6ClientIpNotCallingLimiter() throws Exception {
-    rateLimiterClient.checkEqLaunchLimit(domain, "2001:DB8::21f:5bff:febf:ce22:8a2e", 10);
+    rateLimiterClient.checkWebformRateLimit(domain, "2001:DB8::21f:5bff:febf:ce22:8a2e");
     verifyEnvoyLimiterNotCalled();
   }
 
   @Test
-  public void shouldAcceptRateLimitBelowThreshold_modulo3() throws Exception {
-    doCheckAndVerifyModulo("124.125.126.123", 10, 3);
-  }
+  public void checkWebformRateLimit_belowThreshold() throws CTPException {
+    // Don't need to mock the call to restClient.postResource() as default is treated as being below
+    // the limit
 
-  @Test
-  public void shouldAcceptRateLimitBelowThreshold_modulo4() throws Exception {
-    doCheckAndVerifyModulo("124.125.126.9", 5, 4);
-  }
+    // Run test
+    rateLimiterClient.checkWebformRateLimit(domain, AN_IPv4_ADDRESS);
 
-  @Test
-  public void shouldAcceptRateLimitBelowThreshold_modulo0() throws Exception {
-    doCheckAndVerifyModulo("124.125.126.100", 10, 0);
-  }
-
-  @Test
-  public void shouldAcceptRateLimitBelowThreshold_modulo9() throws Exception {
-    doCheckAndVerifyModulo("124.125.126.249", 15, 9);
-  }
-
-  @Test
-  public void shouldAcceptRateLimitBelowThreshold_modulo19() throws Exception {
-    doCheckAndVerifyModulo("124.125.126.249", 23, 19);
-  }
-
-  private void doCheckAndVerifyModulo(String ipAddress, int loadSheddingModulus, int expectedModulo)
-      throws Exception {
-    rateLimiterClient.checkEqLaunchLimit(domain, ipAddress, loadSheddingModulus);
-
+    // Grab the request sent to the limiter
     RateLimitRequest request = verifiedRequestSentToLimiter();
 
+    // Verify that the limit request contains a ipAddress based descriptor
     assertEquals(1, request.getDescriptors().size());
-    verifyDescriptor(request, 0, "modulo", "" + expectedModulo);
+    verifyDescriptor(request, 0, "ipAddress", AN_IPv4_ADDRESS);
   }
 
   @Test
-  public void shouldRateLimitAboveThreshold() throws Exception {
+  public void checkWebformRateLimit_aboveThreshold() throws Exception {
     // Limit request is going to fail with exception. This needs to contain a string with the
     // limiters too-many-requests response
     ResponseStatusException failureException = overTheLimitException();
@@ -107,7 +76,7 @@ public class RateLimiterClientEqLaunchTest extends RateLimiterClientTestBase {
 
     // Confirm that limiter request fails with a 429 exception
     try {
-      rateLimiterClient.checkEqLaunchLimit(domain, AN_IPv4_ADDRESS, 10);
+      rateLimiterClient.checkWebformRateLimit(domain, AN_IPv4_ADDRESS);
       fail();
     } catch (ResponseStatusException e) {
       assertEquals(failureException, e);
@@ -117,43 +86,43 @@ public class RateLimiterClientEqLaunchTest extends RateLimiterClientTestBase {
   }
 
   @Test
-  public void shouldQuietlyAcceptOtherRateLimiterError() throws Exception {
+  public void checkWebformRateLimit_limiterOtherError() throws Exception {
     // Limit request is going to fail with exception that simulates an unexpected error from the
     // limiter. ie, http response status is neither an expected 200 or 429
     mockRateLimitException(badRequestException());
 
     // Circuit breaker spots that this isn't a TOO_MANY_REQUESTS HttpStatus failure, so
     // we log an error and allow the limit check to pass. ie, no exception thrown
-    rateLimiterClient.checkEqLaunchLimit(domain, AN_IPv4_ADDRESS, 10);
+    rateLimiterClient.checkWebformRateLimit(domain, AN_IPv4_ADDRESS);
     verifiedRequestSentToLimiter();
   }
 
   @Test
-  public void shouldQuietlyAcceptCorruptedJsonResponseFromRateLimiter() throws Exception {
+  public void checkWebformRateLimit_corruptedLimiterJson() throws Exception {
     // This test simulates an internal error in which the call to the limiter has responded
     // with a 429 but the response JSon has somehow been corrupted
     mockRateLimitException(corruptedJsonException());
 
     // Although the rest client call fails the circuit breaker allows the limit check to pass. ie,
     // no exception thrown
-    rateLimiterClient.checkEqLaunchLimit(domain, AN_IPv4_ADDRESS, 10);
+    rateLimiterClient.checkWebformRateLimit(domain, AN_IPv4_ADDRESS);
     verifiedRequestSentToLimiter();
   }
 
   @Test
-  public void shouldWorksWithCircuitBreakerOpen() throws Exception {
+  public void checkWebformRateLimit_worksWithCircuitBreakerOpen() throws Exception {
     // Simulate circuit breaker not calling rest client
     mockRateLimitException(circuitBreakerOpenException);
 
     // Limit check works without an exception
-    rateLimiterClient.checkEqLaunchLimit(domain, AN_IPv4_ADDRESS, 10);
+    rateLimiterClient.checkWebformRateLimit(domain, AN_IPv4_ADDRESS);
   }
 
   private void verifyDescriptor(
       RateLimitRequest request, int index, String finalKeyName, String finalKeyValue) {
     LimitDescriptor descriptor = request.getDescriptors().get(index);
     assertEquals(2, descriptor.getEntries().size());
-    verifyEntry(descriptor, 0, "request", "EQLAUNCH");
+    verifyEntry(descriptor, 0, "request", "WEBFORM");
     verifyEntry(descriptor, 1, finalKeyName, finalKeyValue);
   }
 }
