+++ src/test/java/uk/gov/ons/ctp/integration/ratelimiter/client/RateLimiterClient_IT.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.ctp.integration.ratelimiterclient;
+package uk.gov.ons.ctp.integration.ratelimiter.client;
 
 import static org.junit.Assert.assertEquals;
 
@@ -24,7 +24,6 @@ import uk.gov.ons.ctp.common.rest.RestClientConfig;
 import uk.gov.ons.ctp.integration.common.product.model.Product;
 import uk.gov.ons.ctp.integration.common.product.model.Product.DeliveryChannel;
 import uk.gov.ons.ctp.integration.common.product.model.Product.ProductGroup;
-import uk.gov.ons.ctp.integration.ratelimiter.client.RateLimiterClient;
 
 /**
  * This is a program for manually testing the client against an real/fake limiter.
